requirements.txt
langchain==0.3.26
flask==3.1.1
sentence-transformers==4.1.0
pypdf==5.6.1
python-dotenv==1.1.0
langchain-community==0.3.26
streamlit==1.37.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.7.4
chromadb==0.4.15
langchain-chroma==0.1.4
langchain-groq==0.2.0

-e . 